Match type of arguments to min() and max()
max() and min() were changed from macros to templates in ESP8266 core for Arduino 2.4.0. This change causes calls to these functions with arguments of different types to cause a compilation error "error: no matching function for call to". This commit changes the arguments to these functions to the same type to fix compatibility with ESP8266.

+++ examples/ConsoleShell/ConsoleShell.ino
@@ -140,7 +140,7 @@ void SetLedState()
 void SetBrightness() 
 {
   // clamp value intervalOn on 0 and PWMinterval
-  intervalOn  =  max(min(ledBrightness,PWMinterval),0);
+  intervalOn  =  max(min((unsigned long)ledBrightness,PWMinterval),0ul);
 }
 
 // Pulse Width Modulation to vary Led intensity
